fix: run python_multiprocessing_lock_works integration test on Mac and Linux
